added new logo and replaced it everywhere needed
